Moving interface to it's own file
